♻️ [Refactor] Handler 전체 리팩토링
